🚀 refactor: Service Layer 제거 및 사용하지 않는 생성자 제거
