+++ src/main/java/racingcar/Application.java
@@ -3,15 +3,15 @@ package racingcar;
 
 import racingcar.controller.RacingController;
 import racingcar.domain.Cars;
-import racingcar.service.RacingService;
+import racingcar.domain.RacingGame;
 import racingcar.view.InputView;
 import racingcar.view.ResultView;
 
 public class Application {
 
     public static void main(String[] args) {
-        RacingService racingService = RacingService.from(Cars.instance());
-        RacingController racingController = new RacingController(racingService,
+        RacingGame racingGame = RacingGame.from(Cars.instance());
+        RacingController racingController = new RacingController(racingGame,
             InputView.instance(), ResultView.instance());
         racingController.setUpRacing();
         racingController.startRacing();
