+++ src/test/java/racingcar/domain/MovingTest.java
@@ -3,7 +3,6 @@ package racingcar.domain;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.movable.FixedForwardStrategy;
 import racingcar.domain.movable.RandomForwardStrategy;
-import racingcar.service.RacingService;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -17,14 +16,14 @@ public class MovingTest {
     @Test
     void moveForwardTest() {
         Cars cars = Cars.of(inputs, new FixedForwardStrategy(true));
-        RacingService.from(cars).drive();
+        RacingGame.from(cars).drive();
         assertThat(cars.getCars().get(0).getStep()).isEqualTo(1);
     }
 
     @Test
     void noMoveForwardTest() {
         Cars cars = Cars.of(inputs, new FixedForwardStrategy(false));
-        RacingService.from(cars).drive();
+        RacingGame.from(cars).drive();
         assertThat(cars.getCars().get(1).getStep()).isEqualTo(0);
     }
