+++ src/test/java/racingcar/domain/WinnerTest.java
@@ -3,7 +3,6 @@ package racingcar.domain;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.movable.FixedForwardStrategy;
-import racingcar.service.RacingService;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -12,23 +11,23 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.*;
 
 class WinnerTest {
-    RacingService racingService;
+    RacingGame racingGame;
 
     @BeforeEach
     void setUp() {
         List<String> carNames = new ArrayList<>(Arrays.asList("c", "java", "jpa"));
-        racingService = RacingService.from(Cars.of(carNames, new FixedForwardStrategy(true)));
+        racingGame = RacingGame.from(Cars.of(carNames, new FixedForwardStrategy(true)));
     }
 
     @Test
     void winnerTest() {
-        racingService.drive();
-        assertThat(racingService.judgeWinners().getResult()).containsExactly("c", "java", "jpa");
+        racingGame.drive();
+        assertThat(racingGame.judgeWinners().getResult()).containsExactly("c", "java", "jpa");
     }
 
     @Test
     void noWinnerTest() {
-        assertThat(racingService.judgeWinners().getResult().size()).isEqualTo(0);
+        assertThat(racingGame.judgeWinners().getResult().size()).isEqualTo(0);
     }
 
 }
